+++ docs/source/changelog.rst
@@ -36,6 +36,7 @@ jsonyx 2.0.0 (unreleased)
   :func:`jsonyx.loads` accepts non-ascii digits
 - Improved documentation
 - Improved error messages
+- Made :class:`decimal.Decimal` not JSON serializable by default
 - Made :class:`tuple` JSON serializable
 - Merged ``item_separator`` and ``key_separator`` as ``separators`` for
   :class:`jsonyx.Encoder`, :func:`jsonyx.dump`, :func:`jsonyx.dumps` and
@@ -46,7 +47,6 @@ jsonyx 2.0.0 (unreleased)
 - Removed :data:`!jsonyx.DuplicateKey`
 - Removed :mod:`!jsonyx.tool`
 - Renamed ``python -m jsonyx`` to ``python -m jsonyx format``
-- Sped up decimal encoding
 - Use cache for indentations in the JSON encoder
 
 jsonyx 1.2.1 (Aug 3, 2024)